fix(test): initialize managed env without include
The setup in `managed_can_include_managed` used to push the `composer` environment
when this already included a locally referenced dependency environment.
Pushing environments with local references is no longer allowed so the setup was changed to
create a managed environment without included environments
and add the include separately via an edit operation.