refactor: rename the no-store option, make public
This option is renamed to `--metadata-only` and allows a user to publish
the metadata of a package without uploading the artifact itself. This is
the same behavior as before the rename, but it is now explicitly public
so that a user doesn't need to provide a signing key. The change that
makes the signing key optional is forethcoming.