feat: error with ingress uri and no signing key
This change makes the store URL and signing key options independent.
The signing key argument is now only required under these conditions:
- The custom catalog has a store configured e.g. the `ingress_uri` is
  populated in the response.
- The user has not passed the `--metadata-only` option.

It is a hard error if the user attempts to publish to a catalog with a
configured store, and has not passed the `--metadata-only` flag.

This change also makes the store-url field optional and hides it. This
field will only be used for testing purposes going forward.