feat: refactor modules to fix bootstrap, add README.md
The NixOS/Darwin/Home manager modules are responsible for configuring Nix
substituter and trusted signature settings and for installing Flox on the
system. The problem is that doing both of these at the same time results
in a standalone _build_ of Flox because the binary substituter settings
are not active until after Flox has been built and installed.

This patch refactors the modules to split out common functionality into
a new modules/flox.nix file, and updates each of the respective modules
to set the Nix substituter and trusted signature settings independently
(and irrespective) of the option to install Flox on the system. This in
turn allows the user to install the module, and then enable Flox on a
second pass, thus allowing the download of the pre-built Flox package
to succeed.

This patch also adds a top-level README.md for guiding users through
the process of installing the module, but rather than having them enable
Flox on a second pass we provide the exact `nixos-rebuild switch` option
arguments to be able to exercise the cache.flox.dev substituter without
having to perform a second `nixos-rebuild switch`.

Co-authored-by: Rok Garbas <rok@flox.dev>
Co-authored-by: Thomas Bereknyei <tom@flox.dev>