fix: don't panic trying to install package
If a user publishes a package with metadata only, and another user tries
to install the package, it will attempt to build the package from
source. This hits the code path that in the future will attempt to build
the package from source. That's not supported today, so previously we
would panic. Now we show an error message.