refactor: add existing_lockfile method
It's a common operation to get the lockfile if one already exists. This
method has already existed on CoreEnvironment and this change simply
exposes it on the Environment trait. The new method is used in the
`flox install` command to check the composition override warnings before
and after an installation.