test(managed): Update mocks to include 2 packages
There are two packages present when we run `edit --sync` but because we
don't currently validate catalog responses (issue 1659) we were silently
unlocking "vim" and then the final `uninstall` was removing it and
locking with the already resolved "hello".

This test started failing on the last `uninstall` command when I added
implicit locking to the beginning of `CoreEnvironment::uninstall` in
order to check packages from composition:

    thread 'main' panicked at flox-rust-sdk/src/providers/catalog.rs:866:18:
    expected resolve response, found None

Prevent this from happening by using a mock that has two packages. I've
changed the package names because we already had a suitable curl+hello
mock.

I've created issue 2892 to improve this by explicitly blocking
imperative commands if the lock file is out-of-date.