refactor: consolidate merge_cache_options logic
Consolidate logic for handling cache options instead of splitting it
between publish() and merge_cache_options().

This is more readable, allows unit testing all related logic at the same
time, and skips an unnecessary request to the catalog server when
--store-url is passed.

A test case was added to very that the error for
```
if catalog_has_ingress_uri && no_key_supplied && !metadata_only
```
was not regressed.