feat: heuristics for publishing git remote
Previously we required that the build repository be tracking a remote
branch, where the build repository and the remote repsository were both
on the latest revision. This is restrictive, and causes frustration when
running the `flox publish` command.

This change loosens those restrictions by additionally allowing the
build repository be on a revision that is present on a remote. When
multiple remotes are present we apply the following heuristics to choose
the remote URL to publish. The order of preference for choosing the
remote is as follows:
- The remote whose branch the build repo is tracking
- The remote that is configured if only one is configured
- If multiple remotes are configured, prefer one named "upstream", then
  one named "origin". If none of the remotes have those names, simply
  choose the first one listed by `git remote`.