fix(edit): Ensure lock for noop edits
Change the behaviour of `flox edit` so that it locks and builds an
environment when there are no changes to the manifest and it's not
already locked.

The primary motivation for this is so that we can provide a single
command for users to run in the error message when a composed
environment has unlocked includes. The alternatives are much more janky,
such as:

- `flox edit` with a whitespace change
- `flox activate -- true`

I think it also provides a more consistent experience and better
feedback loop because if you're editing an unlocked environment then
you'll want to know that it builds rather than deferring to the next
time it's activated.

For this reason I think it's also better irrespective of whether we
choose to implement `flox edit --sync` for local environments in:

- https://github.com/flox/flox/issues/2892

Prior to this change the new test would have failed with:

    -- output differs --
    expected : ✅ Environment successfully updated.
    actual   : ⚠️  No changes made to environment.
    --