test(ci): Replace tailscale action with nixpkgs
To workaround the frequent failures we're seeing:

    Downloading https://pkgs.tailscale.com/stable/tailscale_1.82.0_amd64.tgz
    …
     47 30.0M   47 14.3M    0     0   182k      0  0:02:48  0:01:20  0:01:28 89820
    curl: (92) HTTP/2 stream 1 was not closed cleanly: INTERNAL_ERROR (err 2)

There's an upstream issue for this problem and a few PRs to introduce
caching which would help but they haven't had any progress yet:

- https://github.com/tailscale/github-action

So for the timebeing try switching to running from nixpkgs where the
distribution should be more reliable because we'll benefit from both Nix
caching in the workflow and public substitutors.

I've copied the upstream script with some minor modifications and kept
all of our existing arguments. I'm not sure the timeout options make
sense but we can address that separately.