fix: Cannot allocate memory in static TLS block
The runtime loader reserves a static amount of memory for thread local
storage (TLS).
When using LD_AUDIT in combination with libraries such as jemalloc that use TLS,
the default reservation is exceeded and the program fails to start.
Use GLIBC_TUNABLES to increase the reservation.
The default reservation is 512.
Requirements for some programs we've run into the issue with:
On x86_64-linux:
redis-server: 1441
biome: 1417
fd: 1417
On aarch64-linux:
redis-server: 1777
biome: 1553
fd: 1561
See https://sourceware.org/bugzilla/show_bug.cgi?id=31991

Set the reservation to 2500 to give us a little bit of headroom.