refactor: determine ingress_uri in publish provider
It's been decided that the publish provider should communicate with the
catalog about where to publish a package, which includes determining the
store type (which could be meta-only in the near future) and the
ingress_uri.

Move the logic in the CLI and `merge_cache_options` into
`PublishProvider::publish` and `determine_cache`.

This meant that cache was removed from the `PublishProvider` struct.
This makes sense, since PublishProvider is now responsible for
determining if a cache should even be used.

I also dropped the generic `<Cache>` on `impl Publisher for
PublishProvider`. Currently it's allowing using a `MockCache` for a
single test that makes sure that if `upload()` returns an error,
`publish()` returns that error. Since that's essentially testing that
we're using `?`, I don't think it's worth the complexity. It also
doesn't make sense with this refactor to mock at that level - we may
need to test that `publish()` correctly decides which type of cache to
use, and then calls `upload()` with that cache, but completely overriding
the decision of what cache to use just sidesteps all the logic we need
to test.

I also dropped `test_merge_cache_options_success` which has multiple cases
in favor of two existing publish tests and a third test I added
`publish_errors_without_key`. I don't think it's worth unit testing
`determine_cache` as it's just an if + a match, and I don't think at
this point it's worth adding much broader tests that have to call
Publish::handle or something to test every single case.