feat: use /publish endpoint as intended
We've decided that the `/publish` endpoint should be used as a required
step in the publish process. We want to error if a catalog has an
unconfigured store, and we want to use the `/publish` endpoint to get
info required to carry out the publish. For now we need `ingress_uri`,
but we'll likely need `egress_uri` soon as well.

Catalogs are created implicitly server sided when they're needed, so we
can drop `create_catalog` in our client. And we can drop
`get_ingress_uri` in our client in favor of a method named `publish`
since that more clearly represents the underlying endpoint and how we
intend to use it in the future.

We're deliberately changing behavior such that even if `--metadata-only`
is passed, we error for an unconfigured store. We expect users to
configure their catalog for meta-only publishes rather than using
`--metadata-only`, so that option is hidden for now. Since we don't
currently support building from source, if a catalog is configured for
NixCopy but someone publishes with `--metadata-only`, there won't be any
way to build the required store paths.