feat(ci): add more context to BuildKite uploads
Apparently, BuildKite allows us to set custom tags and filter on them. So, I
thought It'd be a good idea to at least include the architecture the suite
targeted, and the Nix falkeref and attribute. E.g.

    -F "tags[architecture]=$MATRIX_SYSTEM" \
    -F "tags[nix_flakeref]=github:flox/flox/$GITHUB_SHA" \
    -F "tags[nix_attribute]=packages.$MATRIX_SYSTEM.flox-cli-tests" \

These tags will then be visible in BuildKite, allowing us to reconstruct the
original command that was ran to execute the tests. In hopes to reproduce an
error if we encounter one.

Ref: https://buildkite.com/docs/test-engine/importing-junit-xml