fix: test_open_subdirectory_bare() GitCommandProvider update
The test_open_subdirectory_bare() test previously considered the existence
of a "branches" subdirectory as confirmation of having successfully created
a bare repository, but with the latest version of the GitCommandProvider it
seems that directory is no longer created. This commit updates the test to
instead look for the "hooks" directory, which it seems continues to be
created with the latest version.