feat: check remote for current revision
Previously during publish we would only check whether the current
revision was on a local copy of a remote branch. Now we actually check
the remote for the revision using the `git fetch --negotiate-only`
command.

This command returns ancestor revisions in common with the local repo
and the remote, and you can limit its scope by providing a
`--negotiation-tip=<rev>` flag. If the output of the command isn't the
current revision, you know that the current revision isn't on the
remote. This command doesn't actually transfer any files.