fix(ci): Remove control characters from XML test report before uploading
Apparently XML files can only contain a certain set of characters, of which some
control characters are not included. We appear to have one (or multiple, I
honestly don't know) of these control characters in the output of our CLI and
bats test suite. Bats doesn't seem to clear out these characters, so we'll just
do it in an extra step.

Ref: https://www.w3.org/TR/REC-xml/#charsets