test: publishing with an unconfigured store
Fix tests broken by recent changes.

Add a test case that checks that we error when trying to publish to a
catalog with an unconfigured store. We should error even if
metadata_only and ingress_uri_override are passed.

I'm not convinced yet that it's the right approach, but I used httpmock
for this test since I wanted to test shim code that converts from an
ErrorResponse to an UnconfiguredCatalog error, and I didn't have a good
way to do that with mocking out the shim completely. We may need to
change this test if we decide we don't want to use httpmock, but I
wanted to try it out.