feat(compose): Support includes of remote envs
Tests are relatively light because most of the supporting work for this
was already done, so I've focused on the fetcher and the happy path for
`flox include upgrade`.

Similarly not much of the documentation needed updating because the
behaviour as described is the same for remote environments. I haven't
duplicated the `manifest.toml` descriptor docs because `name` is in both
and is the most wordy.

Follow-ups that don't need to be in-scope for initial release:

- Locking specific generations: 2948
- Private remotes: 2949

I considered using `env.lockfile().manifest` instead of `env.manifest()`
in the new `fetch_remote()` and existing `fetch_local()` but it doesn't
seem right that these could possibly lock if there's a race condition
between checking the lockfile is up-to-date and getting the lockfile.

Incidentally I'm growing less sure that 6828a500 was a good idea.