feat: pass server error through for /publish
Allow the server to control error details when /publish errors.

I also changed the printing of catalog errors for publish.

Currently for an error we don't pass through any context:

```
$ FLOX_CATALOG_URL="https://example.com" flox publish
...
❌ ERROR: Failed to publish package: There was an error communicating with the catalog
```

After the change we give more useful information:
```
$ FLOX_CATALOG_URL="https://example.com" flox publish
...
❌ ERROR: Failed to publish package: 403 Forbidden
```

For an error on the /publish endpoint, we pass through details:
```
❌ ERROR: Failed to publish package: 422 Unprocessable Entity: Catalog 'mkenigs' must be configured for whether to upload build artifacts.

To configure the catalog to skip uploading artifacts, run:
  $ catalog-util store --catalog mkenigs set meta-only

To configure the catalog to upload artifacts, run:
  $ catalog-util store --catalog mkenigs set nixcopy --ingress-uri <uri> --egress-uri <uri>
```