feat(compose): Remote can include remotes
I misunderstood the intention of this condition and error when
refactoring out `fetch_local()` because remote environments should be
able to include other remote environments.

They can't include local environments though, as covered by an existing
test in `models::environment::RemoteEnvironment`, because we wouldn't be
able to resolve the environment from another location.