fix(compose): Use lockfile for services
So that services are found from the merged manifest.

Before:

    % flox services status -d ~/demo/composed
    ❌ ERROR: Environment does not have any services defined.

After:

    % flox services status -d ~/demo/composed
    NAME       STATUS       PID
    hello      Stopped

Not adding an explicit test for composition with services because it
requires a lot of boilerplate and we're unlikely to regress this
particular case.

I'm going to look into (or ticket if I run out of time) ways to prevent
us from using the unlocked typed manifest, independent from us making a
decision about changing the user facing behaviour in:

- https://github.com/flox/flox/issues/2892