fix(publish): Lock early and use locked manifest
For consistency with what's being built and is committed to the repo.
Will also ensure that it works with merged manifests.

In preparation for removing `Environment::manifest()` so that we stop
accidentally using unlocked and unmerged manifests.

I experimented with passing the lockfile and required paths all the way
through the provider, to make it clearer that we don't need to mutate an
environment, but it affects the signatures in all of the tests so I've
aborted for now.