refactor: Add `BuildEnvNix::gc_root_base_path` field
Our goal is to avoid garbage collection of store paths between realisation
and the eventual nix build producing a flox environment path.
To that end we are going to create gc-roots for all paths going into the environment build.

Since all realisation steps are going to produce gc-roots,
the base paths for gc roots is ubiquitously needed.
In order to avoid passing paths as arguments around, we can use the ambient `BuildEnvNix` instance
to hold a basepath and provide further functionality to format prefixes.