feat(activate): Check trust for remote includes
Check for trust when running `flox activate` against a composed
environment that includes remote environments.

This mimics the behaviour of `flox activate -r` for a non-composed
environment and reduces (or shifts responsibility for) the risk of
running arbitrary hooks from external sources by requiring the user to
review and accept them first.

It's also easier to implement it at `activate` rather than when the
remotes are fetched and merged because that happens deep within the SDK
for several callpaths and we don't have an easy way to surface that
messaging in the CLI.

This check has to be done after locking because, although we have the
names of the remotes from the unlocked manifest, we don't have the
manifest contents of the remotes. This means that the environment will
still be built locally, but we're considering the risk of that being
misused as low compared to running hooks.

I don't like using `assert_output --partial` in tests but the fake
FloxHub generates a bunch of output that makes the tests noisy and I'd
prefer to match the existing style for now.