feat: create .flox/.gitignore for for pull --copy
Previously if you ran a `flox pull --copy` you would not get the
`.flox/.gitignore` file that you get if you run `flox init`. This meant
that logs from services and the watchdog would not be ignored in
environments created via `flox pull --copy`. Now we generate that file
and the logs are ignored.