fix: correctly set pip.ini path in builds
Previously the Python etc/profile.d script would refer to the location
of the `pip.ini` file (necessary to enforce a virtual environment) via
a path relative to `$FLOX_ENV`. This doesn't work in builds because the
relative path is in the Nix store e.g. `/nix/pip.ini`.

This change refactors the `/etc/profile.d` script and the build script
to set and use the `$FLOX_ENV_PROJECT` variable instead of a relative
path.

I tested this interactively but didn't create a test for it because it
didn't seem worth it. To test it I created an environment with a simple
Python project created via Claude. In the build logs you can see that it
correctly creates binary wheels from the project.