fix: remove nix substituter settings from flake
The presence of nix substituter settings always elicits errors for users that have not added themselves as trusted users:
```
flox [limeytexan/default] Michaels-MBP% nix build
warning: ignoring the client-specified setting 'trusted-public-keys', because it is a restricted setting and you are not a trusted user
```

Substituter settings are handled on a system-wide basis by both the Flox installer and the (relatively new) Nix modules so their presence in the flake is unnecessary for anyone who has installed Flox, and presumably everyone developing Flox will have already done that.

Most importantly, we do not want to give the impression that users ought to be adding themselves as trusted users, because that is a well-known way of obtaining root on a system. Our developers should also avoid run as a trusted user so that they will observe / be exposed to the same experience that potential contributers will see when building Flox.