fix: flaky daemon service test
This test was flaking because for some reason this test would take
longer than 2 seconds to run `overmind status`. This was primarily on
x86_64-darwin, but did happen on other systems from time to time. I've
replaced overmind with a daemonized sleep command to try to cut out as
much logic as possible.