feat(publish): FloxHub auth on nix-copy to ingress
Pass a FloxHub token on `nix copy` calls to known FloxHub authenticated
store hostnames. These hostnames are currently hardcoded whilst we
figure out how we will provide multi-tenancy.

I've provided a helper function and then passed the file into the
publisher, rather than passing the token in, so that we don't have to
thread `flox.temp_dir` all the way through and it has symmetry with the
existing signing key argument.

It isn't wrapped in an `Option<>` because it's subject to change (as
above), may cause mistakes if it's sometimes not passed, and has no
effect for non-FloxHub store URIs.

Support for reading narinfo from the egress URI with auth will follow in
a separate commit.