feat: render lockfiles as read-only
This is mostly a preventative measure. Users can see the lockfile right
next to the manifest in their editor, and being able to edit the
manifest in an IDE (outside of the flox edit cycle) was a requested
feature. This is meant to prevent a user from editing the lockfile and
putting their environment into a weird state.

This is accomplished by setting the lockfile to read-only at the end of
a transaction. When we make a writable copy of the environment during
the transaction we set the lockfile back to a writable state by setting
its permissions equal to those of the manifest.