fix: run watchdog for `flox build` activations
We want consistent behavior between building within a flox activation
and running `flox build`, so `flox build` should attach in the same way
that an activation does.
In other words, `flox activate && make` should only run hooks if they
haven't already run, and `flox build` should do the same.

We could introduce a "don't run start or attach" mode in the activation
scripts, but that would require adding an aditional case to the
activation scripts. We can't just skip start or attach when the watchdog
is unset, because we want start or attach to run in containers, which
don't have a watchdog.