feat: add auth provider to buildenv
This adds the `AuthError` type and an `AuthProvider` trait so that we
can mock this with a dummy provider at some point if we want to. At some
point we will need to come up with a solution for bubbling up auth
errors that arise in buildenv and publish. Right now we require a token
for publish, which makes sense, and since we require it we can check for
it early. That's not really the case for buildenv since you don't need
to be logged in just to build an environment.