feat: use auth provider in buildenv
The tricky part in buildenv is that we don't know which stores need auth
beforehand aside from two hardcoded cases:
- `https://cache.nixos.org`
- `daemon`

I don't *really* understand the second one, but Tom claims that it's a
valid value, so I've hardcoded that as well (some of our tests use it,
but without comments on those tests it's hard to tell if those are just
test values or something real).

The existing logic does a pre-build check to make sure that all of the
store paths required to build the environment are substitutable.
Different store paths may be on different stores, so we iterate over
the store paths and the list of stores they can be found in just in case
one store isn't available and another is.

The problem comes from the fact that with build and publish, some stores
require authentication. If you need an auth token and don't have one,
that's an error *for this store*, but there might be another store that
doesn't require auth. So, this adds some complexity to handle that we
don't want to look up an auth token until we have to, and after that
we don't want to error immediately if the token is missing.