refactor: drop `flox` argument to `FloxBuildMk::{build,clean}`
Drop the common `flox` argument to `build` and `clean` methods, to avoid constructing inputs inline in the methodd impl.
Makes the the trait less dependent on the central `Flox` instance
and reduces the number of arguments (particularly to build, where clippy started warning about too many args).

We should further investigare if sharing `base_dir`, `built_environments` and `expression_dir` should be derived upfront too and/or stored on the instance,
as they share the same semantics in both methods.

refactor: store dirs on `FloxBuildMk`

refactor: store verbosity on `FloxBuildMk`