chore(catalog): httpmock for read/write fixtures
Re-purpose the existing environment variables that are used by `mk-data`
and `bats` to read and write catalog mocks to fixture files, replacing
our custom mock generation with `httpmock` YAML files.

This will allow us to consolidate on a single approach throughout all of
our tests. `MockClient` will be removed or replaced later on.

It's worth noting that the mock fixtures are now at the HTTP layer
rather than the shim layer which ensures that we test more of the
codepath and resolves the issues where we don't validate that the mock
response of the shim matches the type of content that we expect, e.g.
using the wrong resolve response or something that's not a resolve
response at all.