test: includes work for manifest builds
Test building a C program that uses an `#include` of a library
provided by Flox.

Tests correspond to the issue described in
https://github.com/flox/flox/issues/2992,
but the reproducer in that issue was technically not a valid reproducer
since in the example given, boost is actually needed at both buildtime
and runtime.
quotes-app-cpp has since been changed accordingly in
https://github.com/flox/flox-manifest-build-examples/pull/6

This commit adds two tests: one that requires boost just at build time,
and one that requires boost at both build time and runtime.

The test that only requires boost at build time fails prior to
5a8e106b30e86ed0e5d8e668f860157e7e81d3d0.
The other one does not, because adding boost to runtime-packages means
that the include of boost can be found via the build-wrapper closure
even though it isn't available via the full development environment
closure.

I originally added a test with zlib, but zlib is present on macOS by
default so we'd have to make assertions about which zlib library was
used, which seems harder than just using boost.