fix: unable to create gc-roots directory
When running `flox list -r` we create a remote environment in a
temporary directory, or pull the latest updates to it if it already
exists. In the past the `.flox/run` path was a symlink to what is now
the `dev` mode environment. That was changed with v1.3.3, and now that
path is a directory that contains the `*.dev` and `*.run` symlinks.
This command attempts to migrate from the old path to the new path,
but failed to do so because of a misunderstanding of the `Path::exists`
method.

The `Path::exists` method traverses symlinks to determine whether the
real file exists. On my system the `.flox/run` symlink was dangling,
so `Path::exists` returned false even though the symlink itself existed.
This caused the `create_dir_all` function to fail to create the new
`.flox/run` directory. We now use the `Path::is_symlink` method to
check whether the symlink exists because it returns true if the path
exists and is a symlink (without traversing it).