chore(mk_data): structural changes to mock gen
The `httpmock` fixtures are YAML with our request and response bodies
nested as JSON, so we have to do a little extra work to manipulate or
read the contents.

We also remove the `Content-Length` header after modifying the body
otherwise the client will complain about the mismatch. We could generate
a new value but it doesn't seem worth the effort right now.

Writing the modified body introduces some linebreaks in the JSON
embedded YAML but these are non-semantic in that they don't affect the
content when deserialized from YAML.