chore(init): auto-setup resolve for all systems
So that request mocks match when run on Darwin and Linux.

I'm in two minds about whether this is good or bad. It could mean that a
project is auto-setup but doesn't actually work on the user's current
system because a package isn't available for that plat/arch. However it
matches the behaviour of `flox install` and the subsequent resolution
request we make for auto-setup.