fix(tests): Unset mock for attach:previous-release
This fails because the older version of Flox uses JSON mocks and the
current test directory that it's operating from only has YAML mocks.

We don't strictly need a mock for the `activate` because the project
doesn't have any packages but we should reinstate it after releasing a
version of Flox that uses YAML mocks because it prevents us accidentally
hitting the real catalog server in tests.