fix(mk-data): Catch errors in cmds
Catch errors in all commands of `pre_cmd`, `cmd`, and `post_cmd` scripts
when `ignore_errors` is not set. Previously only the exit code of the
last command was used to detect errors.

This makes `RESPONSE_FILE` manipulation safer because we won't
accidentally end up with an empty file or field if there's a problem
with a `jq` query.

I tested running `just gen-data -f` locally which completed and verified
that we don't have any current errors. I haven't included it in this
commit though because it's very noisy and will conflict with the
`httpmock` work.