fix(tests): Use single package for edit --sync
So that we don't have to provide mocks for a partially resolved
lockfile now that `httpmock` is being more strict about matching the
request body. Any semantic change to the manifest is sufficient.

Additionally we are more specific about the failure that we're matching
so that the right assertion is triggered when not matching mocks.