chore(tests): Delete extra upgrade by group test
This test requires an new `httpmock` fixture with both the package group
renamed and the `hello` package manipulated to an older version.

It doesn't seem worth the effort of duplicating and generating that
fixture when the `flox upgrade toplevel` test already covers the same
codepath of upgrading a group by name. There isn't any special handling
of the "toplevel" group in the `flox` crate and we have sufficient
coverage further down in the `flox-rust-sdk` crate.