chore(init): All systems for auto-setup resolve
So that request mocks match when run on Darwin and Linux.

I'm in two minds about whether this is approach. It could mean that a
project is auto-setup but doesn't actually work on the user's current
system because a package isn't available for that plat/arch. However it
does match the behaviour of the subsequent resolution request we make
for auto-setup and standard `flox install`. So it seems better than
trying to generate very specific mocks for these cases right now.