refactor: use RawManifest helpers
Rather than re-implementing toml_edit logic in format_customization, add
a RawManifest::new_minimal method that calls RawManifest helpers. It
also avoids doing unnecessary work by calling insert_packages, which is
designed to operate on a pre-existing manifest.

In hindsight I'm not sure it was worth the effort, added conditionals in
RawManifest, and weird newline handling. But it's probably slightly less likely
that we'll end up with accidentally different behavior between new_documented
and format_customization with this approach.