feat: add build support in mk_data
This is distinct from the `envs` group because the layout of the files
is different. The output of each `envs` job tends to be a directory that
contains the original manifest and a lockfile generated from that
manifest. The idea with the `build` section is that you can put an
entire environment in the input data, and the output will also contain
that environment along with the YAML response file e.g. you should be
able to go into the input or output directory and run `flox build`.