refactor: parse empty token as no token
The `mk_data` tools sets the FLOX_FLOXHUB_TOKEN variable to an empty
string to prevent using the host machine's FloxHub token during tests.
However, when the variable is present, its contents are parsed as if
it was a valid token, and an empty string is obviously not a valid
token. This causes `mk_data` to fail.

Instead we can detect an empty string and interpret this as a token not
being present at all. Otherwise, it doesn't seem like there's a way to
tell the CLI to not read the host machine's token.