feat(catalog): Improve semver support
Attempt to parse version specifiers in resolve requests as semver and
normalize them, without any whitespace between multiple comparators, in
order to provide the best possible compatibility with the semver formats
that are supported by the catalog.

There are a few formats which aren't supported by the `semver` library,
which we'll continue to pass on as-is. Some of these are already
supported by the catalog (e.g. `==3.11`, `~=3.11`) and some will be
supported by a subsequent PR that I'll raise of catalog (e.g. `>=3.11
<3.12`, `3.11 - 3.12`). A full list is available here:

- https://docs.google.com/spreadsheets/d/1YfzIjwEGP6GfHoIUrN2DDn8k10BHpkeMnbV84ysle6Y/edit?gid=0#gid=0

I considered adding a `tracing::debug!` when modifying the versions but
it's very noisy when it happens 4x times for each package in each system
group.

I also considered moving this to `Lockfile::collect_package_groups`
where we (mostly) construct `PackageGroup`, which would solve the issue
of tracing once for each install ID, but that means that other resolve
requests like `commands::init::find_compatible_package()` won't benefit
from this behaviour.