test(python): Convert to httpmock
So that we exercise the real catalog resolution every so often. Some of
these fixtures didn't match what they were actually testing because they
were manually written.

Currently most of these tests resolve to the latest version of Python or
Poetry which may need to be bumped when the mocks are re-generated.

I haven't populated an empty response for the tests that don't attempt
resolution because we're planning to change the default test client to
panic if it attempts resolution without being switched out for a
record/replay client.