fix(init): Improve semver support for Python
Don't attempt to pre-parse Python version strings because the `semver`
library that we use in Rust doesn't support some formats that are common
in the Python ecosystem and already supported by the catalog, such as
`==3.10` and `~=3.10`.

The library also rendered `semver::VersionReq` with spaces after the
comma separators which isn't supported by the catalog, such as `>=3.11,
<3.12` and `>= 3.11, < 3.12`.

Now that we're not parsing we need to ensure that any whitespace in the
input is stripped before passing it to the catalog. This appears safe to
do in the case of Python version formats but it wouldn't be safe to do
for all manifests or catalog resolution requests because it would make
NodeJS formats like `3.11 - 3.12` ambiguous.

I was originally hoping to parse all catalog requests on the CLI and
remove the whitespace on the way out but this turned out not to be
possible because `semver::VersionReq` renders `1.2.3` as `^1.2.3` which
changes the meaning to `1.2.x`. It also further splits the logic between
the CLI and the catalog server, which is best kept in one place if we
can help it.