test: add unit test re-generation to just gen-data
Modify just gen-data to additionally run all unit tests in order to regenerate
all mocks for unit tests. This is slower than necessary since it runs all tests,
not just those with mocks, but since we shouldn't be regenerating too often, and
it's essentially the same speed as our unit tests, it should be fast enough for
now.

I ran just gen-data to test and committed the resulting mocks.