fix: return error on build id conflict
When a flox build name was provided by both the manifest via `build.<name>` _and_ as a nix expression in `.flox/pkgs/<name>{.nix/default.nix}` we previously fell back to the expression build printing a warning in the process.
However in other collision situations we anticipate errors being thrown instead,
together with directions to mitigate the error.
For the sake of consistency we'll change the behavior here as well.